+++ source/Private/Get-DpIntercomPayload.ps1
@@ -39,6 +39,7 @@ function Get-DpIntercomPayload {
         chatId                 = [string]$settings.chatId
         allowGroupChat         = [bool]$settings.allowGroupChat
         groupChatIds           = @($settings.groupChatIds)
+        requireGroupMention    = [bool](Get-DpPropertyValue -InputObject $settings -Name @('requireGroupMention') -Default $false)
         heartbeatMinutes       = [int]$settings.heartbeatMinutes
         stallMinutes           = [int]$settings.stallMinutes
         questionTimeoutMinutes = [int]$settings.questionTimeoutMinutes